feat: extract feature flags into @gleanwork/docusaurus-plugin-feature-flags
Moves the bespoke feature flag system out of the site and into a reusable
internal Docusaurus plugin/theme package so the new glean-docs site can
share the same implementation without duplication.

Changes:
- Create packages/docusaurus-plugin-feature-flags/ as a hybrid plugin+theme
  package following the docusaurus-theme-glean pattern
- Move core evaluation logic (featureFlags.ts, featureFlagTypes.ts, tests)
  from src/lib/ into the package's src/lib/
- Move build-time utilities (buildTimeFlags.ts, filtering.ts) from src/utils/
  into the package's src/build/
- Extract FeatureFlagsProvider and its React context from Root.tsx into
  src/theme/FeatureFlagsProvider/, registered automatically via getThemePath()
- Move FeatureFlag component and DocSidebarItems theme override into the package
- Add withFeatureFlags() config wrapper replacing two IIFEs in docusaurus.config.ts
- Update site to consume the plugin via @theme/* aliases (SSR-safe) and
  workspace:* dependency
- Update test mock to reference new package context export path
