+++ docs/guides/mcp/RemoteMCPContent.mdx
@@ -4,7 +4,7 @@ import TabItem from '@theme/TabItem';
 import Admonition from '@theme/Admonition';
 import { CLIENT, MCPConfigRegistry } from '@gleanwork/mcp-config-schema/browser';
 import { clientNeedsMcpRemote } from '@gleanwork/mcp-config-schema';
-import { FeatureFlagsContext } from '@site/src/theme/Root';
+
 
 
 # <Icon name="mcp" iconSet="glean" className="inline" height="1.4rem" /> Model Context Protocol (MCP) Remote Server
