+++ docs/guides/mcp/mcp.mdx
@@ -5,7 +5,6 @@ hide_table_of_contents: false
 ---
 
 import React from 'react';
-import FeatureFlag from '@site/src/components/FeatureFlag';
 import RemoteMCPContent from './RemoteMCPContent.mdx';
 
 <RemoteMCPContent />
