+++ package.json
@@ -92,6 +92,7 @@
     "@docusaurus/utils": "^3.9.2",
     "@docusaurus/utils-validation": "^3.9.2",
     "@gleanwork/api-client": "^0.13.4",
+    "@gleanwork/docusaurus-plugin-feature-flags": "workspace:*",
     "@gleanwork/docusaurus-theme-glean": "workspace:*",
     "@gleanwork/mcp-config-schema": "^4.1.0",
     "@intercom/messenger-js-sdk": "^0.0.18",
