+++ packages/docusaurus-plugin-feature-flags/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@gleanwork/docusaurus-plugin-feature-flags",
+  "version": "0.1.0",
+  "private": true,
+  "main": "src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./context": "./src/theme/FeatureFlagsProvider/context.ts",
+    "./types": "./src/types.ts",
+    "./evaluation": "./src/lib/featureFlags.ts",
+    "./build": "./src/build/buildTimeFlags.ts",
+    "./filtering": "./src/build/filtering.ts",
+    "./withFeatureFlags": "./src/withFeatureFlags.ts",
+    "./FeatureFlag": "./src/theme/FeatureFlag/index.tsx"
+  },
+  "peerDependencies": {
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/plugin-content-docs": "^3.0.0",
+    "react": "^18.0.0"
+  }
+}
