+++ packages/docusaurus-plugin-feature-flags/src/types.ts
@@ -1,0 +1,5 @@
+export type FeatureFlagPluginOptions = {
+  apiEndpoint?: string;
+  cacheTtlMs?: number;
+  debug?: boolean;
+};
