+++ packages/docusaurus-plugin-feature-flags/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@docusaurus/tsconfig",
+  "compilerOptions": {
+    "baseUrl": "."
+  }
+}
